+++ test/services/messenger/services/messengerPermissionService.test.js
@@ -12,6 +12,7 @@ describe('MessengerPermissionService', function test() {
 		configBefore = Configuration.toObject(); // deep copy current config
 		Configuration.set('FEATURE_RABBITMQ_ENABLED', true);
 		Configuration.set('FEATURE_MATRIX_MESSENGER_ENABLED', true);
+		Configuration.set('MATRIX_MESSENGER__STUDENT_ROOM_CREATION', true);
 		// eslint-disable-next-line global-require
 		app = await require('../../../../src/app');
 		// eslint-disable-next-line global-require
@@ -33,7 +34,7 @@ describe('MessengerPermissionService', function test() {
 		params.route = { schoolId: school._id.toString() };
 
 		const result = await app.service('schools/:schoolId/messengerPermissions').get(adminUser._id, params);
-		expect(result.studentsMessengerRoomCreate).to.be.true;
+		expect(result.createRoom).to.be.true;
 	});
 
 	it('teachers have messenger permission to open one to one chats', async () => {
@@ -44,7 +45,7 @@ describe('MessengerPermissionService', function test() {
 		params.route = { schoolId: school._id.toString() };
 
 		const result = await app.service('schools/:schoolId/messengerPermissions').get(teacher._id, params);
-		expect(result.studentsMessengerRoomCreate).to.be.true;
+		expect(result.createRoom).to.be.true;
 	});
 
 	it('students do not have messenger permission to open one to one chats by default', async () => {
@@ -55,7 +56,7 @@ describe('MessengerPermissionService', function test() {
 		params.route = { schoolId: school._id.toString() };
 
 		const result = await app.service('schools/:schoolId/messengerPermissions').get(student._id, params);
-		expect(result.studentsMessengerRoomCreate).to.be.false;
+		expect(result.createRoom).to.be.false;
 	});
 
 	it('students have messenger permission to open one to one chats if school setting is set', async () => {
@@ -66,7 +67,7 @@ describe('MessengerPermissionService', function test() {
 		params.route = { schoolId: school._id.toString() };
 
 		const result = await app.service('schools/:schoolId/messengerPermissions').get(student._id, params);
-		expect(result.studentsMessengerRoomCreate).to.be.true;
+		expect(result.createRoom).to.be.true;
 	});
 
 	it('administrators from other school do not have messenger permission to open one to one chats', async () => {
@@ -78,7 +79,7 @@ describe('MessengerPermissionService', function test() {
 		params.route = { schoolId: school._id.toString() };
 
 		const result = await app.service('schools/:schoolId/messengerPermissions').get(adminUser._id, params);
-		expect(result.studentsMessengerRoomCreate).to.be.false;
+		expect(result.createRoom).to.be.false;
 	});
 
 	it('administrators from other school do not have messenger permission to open one to one chats even if school setting is set', async () => {
@@ -90,6 +91,6 @@ describe('MessengerPermissionService', function test() {
 		params.route = { schoolId: school._id.toString() };
 
 		const result = await app.service('schools/:schoolId/messengerPermissions').get(adminUser._id, params);
-		expect(result.studentsMessengerRoomCreate).to.be.false;
+		expect(result.createRoom).to.be.false;
 	});
 });